expgfx expgfx_addremove: attachedSource NULL-deref bug repro + conditional null web-merge + slot/tex decl-order + zero-init order + pi2-shared tail block + int boundsTemplateId w/ u8 cast (98.03%->98.84%)

+++ include/main/dll_000A_expgfx.h
@@ -29,7 +29,7 @@ void expgfx_resetAllPools(void);
 void expgfx_updateFrameState(int sourceMode,int sourceId);
 struct ExpgfxSpawnConfig;
 int expgfx_addremove(struct ExpgfxSpawnConfig *config, int preferredPoolIndex, int slotType,
-                     u8 boundsTemplateId);
+                     int boundsTemplateId);
 void expgfx_onMapSetup(void);
 void expgfx_release(void);
 void expgfx_initialise(void);
